requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
JPype1==1.5.0
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tabula-py==2.9.0
tornado==6.4
traitlets==5.14.1
tzdata==2023.4
wcwidth==0.2.13
xlrd==2.0.1
